feat(wizard): Add handlers for previousStep and goToStep
This allows users to attach callbacks to be called when navigating backwards,
and jumping in the step hiearchy. This takes the handler for next step as
inspiration for the other triggers, encapsulating them in the same async call methods.

These differ from usig the onStepChange method, and allows for attaching different behaviour
inside step components instead of higher up in a component hiearachy.